target-tracking/requirements.txt
psycopg2-binary==2.8.2
PyMySQL==0.9.3
zappa==0.48.2
